Added the ability to use  Breeze local DB
You can use the interface to save ( and use ) all the data currently
requested from the server.
Saving those data will allow to evade ask the same infos to the server.
In the future the save operation will be handled automatically by
Ancilla.
In the future Ancilla will be able to merge server/client datas to
handle correctly the information coherence